Refactor: 정보뷰 리팩토링 및 정보뷰모델 생성
